feat : AvataToolsContainer 생성 - 아이콘들이 조금씩 추가됨에 따라 container가 필요해짐
