+++ src/components/AvataToolsContainer/DeleteButton/DeleteButton.ts
@@ -1,0 +1,3 @@
+function DeleteButton() {}
+
+export default DeleteButton;
